Refactor: `CommentInput` folder structure
