+++ src/modules/core/components/Comment/Input/CommentInput.css
