+++ src/modules/core/components/Comment/index.ts
@@ -1,1 +1,3 @@
 export { default, Props } from './Comment';
+
+export { default as CommentInput } from './Input';
